Make QueryEditor as a reusable react component
