+++ precise/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/png" href="/commander_bunbun.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Trino Query Editor</title>
+    <title>Trino Query Editor - Example app</title>
   </head>
   <body>
     <div id="root"></div>
