--- precise/src/App.tsx
@@ -1,12 +1,0 @@
-import QueryApp from './QueryApp'
-
-// for now this is a defacto Tab, but we will treat this tab more as a page in the future
-function App() {
-    return (
-        <div className="page">
-            <QueryApp />
-        </div>
-    )
-}
-
-export default App
