+++ precise/src/index.ts
@@ -1,0 +1,1 @@
+export { default as QueryEditor } from './QueryEditor'
