+++ precise/src/main.tsx
@@ -1,12 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
-import './style/theme.css'
-import './style/normalize.css'
-import './style/layout.css'
-import './style/components.css'
-import './style/query-editor.css'
-import './style/results.css'
-import './style/control.css'
+import QueryEditor from './QueryEditor'
 
-ReactDOM.createRoot(document.getElementById('root')!).render(<App />)
+ReactDOM.createRoot(document.getElementById('root')!).render(
+    <React.StrictMode>
+        <h1>Trino Query Editor - Example app</h1>
+        <QueryEditor />
+    </React.StrictMode>
+)
