+++ precise/vite.config.ts
@@ -4,7 +4,6 @@ import react from '@vitejs/plugin-react'
 // Used for local debugging against Trino also running on localhost at port 8080:
 // https://vitejs.dev/config/
 export default defineConfig({
-  base: '/query/',
   plugins: [react()],
   server: {
     proxy: {
@@ -14,6 +13,34 @@ export default defineConfig({
         secure: false,
       },
     },
+  },
+  build: {
+    lib: {
+      entry: "src/index.ts",
+      name: "QueryEditor",
+      fileName: "index"
+    },
+    rollupOptions: {
+      // Don’t bundle peer dependencies like React
+      external: [
+        "react",
+        "react-dom",
+        "react-dom/client",
+        "react-dom/server",
+        "react/jsx-runtime",
+        "react/jsx-dev-runtime"
+      ],
+      output: {
+        globals: {
+          react: "React",
+          "react-dom": "ReactDOM",
+          "react-dom/client": "ReactDOMClient",
+          "react-dom/server": "ReactDOMServer",
+          "react/jsx-runtime": "jsxRuntime",
+          "react/jsx-dev-runtime": "jsxDevRuntime"
+        }
+      }
+    }
   },
 });
